requirements.txt
#trying this out
pandas
numpy
requests
seaborn
plotly.express


